Put electron menu into separate folder
